+++ test/clojucture/core_test.clj
@@ -1,4 +1,3 @@
 (ns clojucture.core-test
   (:require [clojure.test :refer :all]
-            [clojucture.asset :refer :all]
-            [clj-time.core :as t]))
+            [clojucture.asset :refer :all]))