+++ test/clojucture/local_cn.clj
@@ -1,0 +1,2 @@
+(ns clojucture.local-cn
+  (:require [clojure.test :refer :all]))