+++ test/clojucture/pool_test.clj
@@ -1,7 +1,7 @@
-(comment
 (ns clojucture.pool-test
   (:require
     [java-time :as jt]
+    [clojure.test :refer :all]
     [clojucture.asset :as asset ]
     [clojucture.asset-test :as at]
     [clojucture.pool :as p]
@@ -12,11 +12,8 @@
 (def date-intervals
   (u/gen-dates-interval [(jt/local-date 2016 4 1) (jt/local-date 2018 4 1) (jt/local-date 2023 1 2)]))
 
-(def test-pool
-  (p/->pool
-    [ at/test-loan at/test-loan2 at/test-loan3]
-    date-intervals
-    )
-)
 
-)
+(deftest test-pool-cf
+         (let [tp (p/->pool [ ])] )
+
+         )